requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.11.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Markdown==3.3.4
PyJWT==2.3.0
pytz==2021.1
sqlparse==0.4.1
